Refactor: take test to another file
